refactor(App): separate game into a new file
App 클래스레서 작성한 게임 관련 기능을 BaseBallGame.js로 분리하였습니다.
- BaseBallGame 클래스를 추가하였습니다.
  - 이에 따라 중복되는 메서드의 이름을 대폭 수정하였습니다.
  - Before -> After 순서입니다.
  - createNewGame -> create
  - runGame -> run
  - runTurn -> turn
  - quitGame -> quit
- 분리하는 과정에서 다음과 같은 함수가 추가되었습니다.
  - 메시지를 전달받을 경우에만 출력하는 printMessage 함수
